Add Budget editing for act parts

+++ Assets/VC/Piechart/Script/CanvasDataManager.cs
@@ -25,12 +25,12 @@ namespace PieChart.ViitorCloud
 
             }
         }
-        public void PiechartCreated(float sumofdata, float[] data, string[] dataDescription, GameObject[] pieObjects)
+        public void PiechartCreated(decimal sumofdata, decimal[] data, string[] dataDescription, GameObject[] pieObjects)
         {
             for (int i = 0; i < pieObjects.Length; i++)
             {
                 gps[i].transform.GetComponent<Image>().color = pieObjects[i].GetComponent<MeshRenderer>().material.color;
-                gps[i].transform.Find("PercentageText").GetComponent<Text>().text = (Mathf.RoundToInt((data[i] * 100) / sumofdata)).ToString() + "%";
+                gps[i].transform.Find("PercentageText").GetComponent<Text>().text = (Mathf.RoundToInt((float)((data[i] * 100) / sumofdata))).ToString() + "%";
                 //if (dataHeadername.Length > 0)
                 //    gps[i].transform.Find("HeadingText").GetComponent<Text>().text = dataHeadername[i];
                 if (dataDescription.Length > 0)
